Changed to long[] sales

+++ src/ru/koylubaevaem/task2git/Main.java
@@ -5,13 +5,13 @@ public class Main {
 
     public static void main(String[] args) {
 
-        int[] sales = new int[SIZE];
+        long[] sales = new long[SIZE];
         for (int i = 0; i < SIZE; i++) {
-            sales[i] = (int) (Math.random() * 100);
+            sales[i] = (long) (Math.random() * 100);
         }
 
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println("The highest monthly sales " + max);
     }
 }
